pyproject.toml
[project]
name = "bsforge"
version = "0.1.0"
description = "AI-powered YouTube Shorts automation system"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "BSForge Team" }
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "aiosqlite>=0.19.0",
    "redis>=5.0.1",
    # Task Queue
    "celery[redis]>=5.3.4",
    "flower>=2.0.1",
    # LLM & AI
    "anthropic>=0.18.0",
    "litellm>=1.50.0",
    "langchain>=0.1.5",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.0.17",
    "openai>=1.10.0",
    # Embeddings & Vector DB
    "sentence-transformers>=2.3.1",
    "pgvector>=0.2.5",
    "torch>=2.1.0",
    # Web Scraping & API
    "httpx>=0.26.0",
    "scrapy>=2.11.0",
    "playwright>=1.41.0",
    "beautifulsoup4>=4.12.3",
    "feedparser>=6.0.11",
    "praw>=7.7.1",
    # Video Generation
    "edge-tts>=6.1.10",
    "elevenlabs>=0.2.27",
    "pysrt>=1.1.2",
    "pillow>=10.2.0",
    "ffmpeg-python>=0.2.0",
    "yt-dlp>=2024.1.0",
    # YouTube
    "google-api-python-client>=2.115.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    # Dependency Injection
    "dependency-injector>=4.41.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "mako>=1.3.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=24.1.0",
    "watchfiles>=0.21.0",
    # Data Processing
    "pandas>=2.2.0",
    "numpy>=1.26.3",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    # Date & Time
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
    "pytrends>=4.9.2",
    "pycountry>=24.6.1",
    "fontconfig-py>=1.0.0",
    "babel>=2.17.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.26.0",
    "faker>=22.6.0",

    # Code Quality
    "black>=24.1.1",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",

    # Type Stubs
    "types-pyyaml>=6.0.12",
    "types-redis>=4.6.0",
    "types-python-dateutil>=2.8.19",
    "types-requests>=2.31.0",
    "types-beautifulsoup4>=4.12.0",

    # Development Tools
    "ipython>=8.20.0",
    "ipdb>=0.13.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--cov=app",
    "--cov-report=html",
    "--cov-report=term-missing",
    "-v",
    "--strict-markers",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.pyc",
    "migrations",
    "alembic",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "E4",  # pycodestyle import rules (E401 multiple imports on one line, E402 module level import not at top)
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/*" = ["T201", "S101"]  # Allow print and assert in tests
"scripts/*" = ["T201"]  # Allow print in scripts (CLI output)

[tool.ruff.lint.isort]
known-first-party = ["app"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "anthropic.*",
    "litellm.*",
    "langchain.*",
    "pgvector.*",
    "sentence_transformers.*",
    "scrapy.*",
    "playwright.*",
    "praw.*",
    "edge_tts.*",
    "elevenlabs.*",
    "dependency_injector.*",
    "feedparser.*",
    "pytrends.*",
]
ignore_missing_imports = true

# DI Container uses string-based class references which mypy cannot infer
[[tool.mypy.overrides]]
module = "app.core.container"
disable_error_code = ["var-annotated", "no-any-return", "misc", "type-arg", "no-untyped-def"]

# Celery workers use decorators that lose type info
[[tool.mypy.overrides]]
module = "app.workers.*"
disable_error_code = ["no-untyped-def", "no-untyped-call", "untyped-decorator", "attr-defined"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
